requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
boto3==1.26.122
botocore==1.29.122
bracex==2.3.post1
certifi==2022.12.7
charset-normalizer==3.1.0
cli-exit-tools==1.2.3.2
click==8.1.3
colorama==0.4.6
dataclasses-json==0.5.7
deeplake==3.3.2
dill==0.3.6
entrypoints==0.4
frozenlist==1.3.3
greenlet==2.0.2
humbug==0.3.1
idna==3.4
igittigitt==2.1.2
jmespath==1.0.1
langchain==0.0.150
lib-detect-testenv==2.0.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
numcodecs==0.11.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.5
openapi-schema-pydantic==1.2.4
packaging==23.1
pathos==0.3.0
Pillow==9.5.0
pox==0.3.2
ppft==1.7.6.6
pydantic==1.10.7
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.3.23
requests==2.29.0
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==2.0.11
tenacity==8.2.2
tiktoken==0.3.3
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
wcmatch==8.4.1
yarl==1.9.2
